rename ColorTableTypes.ts to ColorMapFunction.ts
