+++ typescript/packages/well-log-viewer/src/components/WellLogView.test.tsx
@@ -13,7 +13,7 @@ const exampleWellLog = {
     curves: [],
     data: [],
 };
-import type { ColorMapFunction } from "./ColorTableTypes";
+import type ColorMapFunction from "./ColorMapFunction";
 const exampleColorMapFunctions = colorTables as ColorMapFunction[];
 
 window.ResizeObserver =
